# fhir-iq-prior-auth/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FHIR IQ Prior Authorization Support API
  description: |
    FHIR-based Prior Authorization Support (PAS) API implementing Da Vinci PAS and DTR
    Implementation Guides for CMS-0057-F compliance demonstration.

    This API provides endpoints for:
    - Submitting prior authorization requests (PAS)
    - Inquiring about authorization status
    - Documentation Templates and Rules (DTR) support
    - Provider and Patient Access APIs
    - Payer-to-payer bulk export operations
  version: 1.0.0-poc
  contact:
    name: FHIR IQ Integration Team
    email: integration@fhir-iq.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://pas-api.fhir-iq.com/fhir
    description: Production FHIR IQ PAS Server
  - url: http://localhost:3000/fhir
    description: Local development server
security:
  - OAuth2:
      - system/Claim.write
      - user/Claim.read
      - patient/Claim.read
paths:
...
  /Questionnaire:
    get:
      summary: Search DTR Questionnaires
      description: Retrieve questionnaires for documentation requirements
      operationId: searchQuestionnaires
      security:
        - OAuth2:
            - user/Questionnaire.read
            - system/Questionnaire.read
      parameters:
        - name: url
          in: query
          description: Canonical URL of the questionnaire
          schema:
            type: string
            format: uri
        - name: context
          in: query
          description: Use context (service type)
          schema:
            type: string
        - name: _count
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Available questionnaires
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Bundle'
  /QuestionnaireResponse:
    post:
      summary: Submit Questionnaire Response
      description: Submit completed DTR questionnaire response
      operationId: createQuestionnaireResponse
      security:
        - OAuth2:
            - user/QuestionnaireResponse.write
            - system/QuestionnaireResponse.write
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/QuestionnaireResponse'
      responses:
        '201':
          description: QuestionnaireResponse created
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/QuestionnaireResponse'
  /Questionnaire/{id}/$populate:
    post:
      summary: Populate Questionnaire with Patient Data
      description: |
        Prepopulate a questionnaire with patient data using CQL expressions
        as defined in the Da Vinci DTR Implementation Guide.
      operationId: populateQuestionnaire
      security:
        - OAuth2:
            - user/Questionnaire.read
            - system/Questionnaire.read
      parameters:
        - name: id
          in: path
          required: true
          description: Questionnaire resource ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              type: object
              properties:
                resourceType:
                  type: string
                  enum: [Parameters]
                parameter:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        enum: [subject, practitioner, organization]
                      valueReference:
                        $ref: '#/components/schemas/Reference'
      responses:
        '200':
          description: Prepopulated questionnaire response
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/QuestionnaireResponse'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.fhir-iq.com/oauth2/authorize
          tokenUrl: https://auth.fhir-iq.com/oauth2/token
          scopes:
            system/Claim.read: Read claim resources (system)
            system/Claim.write: Write claim resources (system)
            user/Claim.read: Read claim resources (user)
            user/ClaimResponse.read: Read claim response resources (user)
            patient/Claim.read: Read patient's claim resources
            patient/ClaimResponse.read: Read patient's claim responses
            user/Questionnaire.read: Read questionnaire resources
            user/QuestionnaireResponse.write: Create questionnaire responses
            system/Patient.read: Read patient resources (system)
            system/export: Perform bulk export operations
        clientCredentials:
          tokenUrl: https://auth.fhir-iq.com/oauth2/token
          scopes:
            system/Claim.write: Write claim resources (system)
            system/export: Perform bulk export operations
  schemas:
    Bundle:
      type: object
      properties:
        resourceType:
          type: string
          enum: [Bundle]
        type:
          type: string
          enum: [document, collection, searchset, batch-response]
        total:
          type: integer
        entry:
          type: array
          items:
            type: object
...
    QuestionnaireResponse:
      type: object
      properties:
        resourceType:
          type: string
          enum: [QuestionnaireResponse]
        status:
          type: string
          enum: [in-progress, completed, amended, entered-in-error, stopped]
        questionnaire:
          type: string
          format: uri
...
    Reference:
      type: object
      properties:
        reference:
          type: string
        display:
          type: string
...